+++ index.js
@@ -1,47 +1,24 @@
-const TelegramBot = require('node-telegram-bot-api');
-const fs = require('fs');
+const Telegraf = require('telegraf')
+const Extra = require('telegraf/extra')
+const Markup = require('telegraf/markup')
 require('dotenv').config();
+
+const bot = new Telegraf(process.env.BOT_TOKEN)
+
+const fs = require('fs');
 
 let greeting = fs.readFileSync('./greeting.txt', { encoding: 'utf-8' });
 
-const bot = new TelegramBot(process.env.BOT_TOKEN, {
-    polling: true,
-    timeout: 500
-});
+bot.on('new_chat_members', (ctx) => {
+    // console.log(ctx.tg)
+    // console.log(ctx.update)
+    const userName = `${ctx.update.message.new_chat_member.first_name || ''} ${ctx.update.message.new_chat_member.last_name || ''}`
+    const text = greeting.replace('{kroshka}', `[${userName}](tg://user?id=${ctx.update.message.new_chat_member.id})`);
+    ctx.replyWithMarkdown(text)
+})
 
-bot.on('message', msg =>{
-    if (msg.new_chat_members) {
-        const new_user = msg.new_chat_member;
 
-        new_user.username = `${new_user.first_name || ''} ${new_user.last_name || ''}`
 
-        const text = greeting.replace('{kroshka}', `[${new_user.username}](tg://user?id=${new_user.id})`);
 
-        bot.sendMessage(msg.chat.id, text, { parse_mode: 'markdown' });
-    }
-});
+bot.launch()    
 
-bot.onText(/^\/change_text (\S+) ([\s\S]*)$/gm, (msg, match) => {
-    const [, password, newText] = match;
-
-    if (password === (process.env.PASS || 'apirol')) {
-        try {
-            fs.writeFileSync('./greeting.txt', newText, { encoding: 'utf-8'});
-
-            greeting = newText;
-
-            bot.sendMessage(msg.chat.id, `Текст изменен на: *${newText}*`);
-        } catch (err) {
-            console.error(err)
-        }
-    }
-});
-
-bot.on('polling_error', console.error);
-
-// hack
-const http = require('http');
-http.createServer((req, res) => {
-    res.write('Hello World!');
-    res.end();
-}).listen(process.env.PORT || 8084);
